move from pipeline utils to gen utils
