+++ evalml/pipelines/time_series_pipeline_base.py
@@ -5,6 +5,7 @@ import woodwork as ww
 from evalml.pipelines import PipelineBase
 from evalml.pipelines.pipeline_meta import PipelineBaseMeta
 from evalml.utils import drop_rows_with_nans, infer_feature_types
+from evalml.utils.gen_utils import are_datasets_separated_by_gap_time_index
 
 
 class TimeSeriesPipelineBase(PipelineBase, metaclass=PipelineBaseMeta):
@@ -71,9 +72,6 @@ class TimeSeriesPipelineBase(PipelineBase, metaclass=PipelineBaseMeta):
 
         Need to do this so that we have all the data we need to compute lagged features on the holdout set.
         """
-        from evalml.pipelines.utils import (
-            are_datasets_separated_by_gap_time_index,
-        )
 
         last_row_of_training = self.forecast_horizon + self.max_delay + self.gap
         gap_features = pd.DataFrame()
